+++ feature/metis/conversation/saved-posts/src/test/java/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/saved_posts/SavedPostsScreenUiTest.kt
@@ -1,4 +1,4 @@
-package de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui
+package de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.saved_posts
 
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.semantics.SemanticsActions
@@ -9,8 +9,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import de.tum.informatics.www1.artemis.native_app.core.common.test.UnitTest
 import de.tum.informatics.www1.artemis.native_app.core.data.DataState
 import de.tum.informatics.www1.artemis.native_app.core.test.BaseComposeTest
-import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.R
-import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.SavedPostsTestUtil
+import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.saved_posts.ui.SavedPostsScreen
 import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto.ISavedPost
 import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto.SavedPostStatus
 import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto.StandalonePost
@@ -107,7 +106,7 @@ class SavedPostsScreenUiTest : BaseComposeTest() {
         onRemoveFromSavedPosts: (ISavedPost) -> Deferred<de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.shared.service.MetisModificationFailure?> = { CompletableDeferred() }
     ) {
         composeTestRule.setContent {
-            de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.saved_posts.ui.SavedPostsScreen(
+            SavedPostsScreen(
                 modifier = Modifier,
                 status = forStatus,
                 savedPostsDataState = DataState.Success(listOf(savedPost)),
