+++ feature/metis/conversation/src/test/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/service/network/impl/SavedPostServiceImplE2eTest.kt
@@ -4,7 +4,7 @@ import de.tum.informatics.www1.artemis.native_app.core.common.test.EndToEndTest
 import de.tum.informatics.www1.artemis.native_app.core.common.test.testServerUrl
 import de.tum.informatics.www1.artemis.native_app.core.test.test_setup.DefaultTimeoutMillis
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ConversationMessagesBaseTest
-import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.SavedPostsTestUtil
+import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.saved_posts.SavedPostsTestUtil
 import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto.SavedPostStatus
 import kotlinx.coroutines.test.runTest
 import org.junit.Ignore
@@ -94,7 +94,7 @@ class SavedPostServiceImplTest : ConversationMessagesBaseTest() {
                 .orThrow("Could not save post")
 
             sut.changeSavedPostStatus(
-                post = SavedPostsTestUtil.fromStandalonePost(post),
+                post = de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.saved_posts.SavedPostsTestUtil.fromStandalonePost(post),
                 status = SavedPostStatus.COMPLETED,
                 authToken = accessToken,
                 serverUrl = testServerUrl
